# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Kivro API
  version: 1.0.0
  description: |
    The Kivro API lets you generate, verify, and track digital addresses 
    across all 54 African countries. Use `/api/v1/` for all new integrations.
  contact:
    name: Kivro Developer Support
    url: https://kivro.africa/help
    email: developers@kivro.africa
  license:
    name: Proprietary
    url: https://kivro.africa/terms
servers:
  - url: https://kivro-fullstack.vercel.app/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass your Kivro API key as `Bearer kv_live_<key>` 
        OR your Supabase JWT for user-authenticated endpoints.
  schemas:
...
    ApiKey:
      type: object
      properties:
        id:               { type: string, format: uuid }
        name:             { type: string }
        key_prefix:       { type: string, example: "kv_live_ab12cd34" }
        plan:             { type: string, enum: [pro, business, enterprise, government] }
        status:           { type: string, enum: [active, revoked] }
        rate_limit_per_min: { type: integer }
        requests_today:   { type: integer }
        created_at:       { type: string, format: date-time }
...
paths:
...
  /keys:
    get:
      summary: List API keys
      tags: [API Keys]
      responses:
        '200':
          description: List of keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys: { type: array, items: { $ref: '#/components/schemas/ApiKey' } }
    post:
      summary: Create an API key
      tags: [API Keys]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, plan]
              properties:
                name: { type: string }
                plan: { type: string, enum: [pro, business, enterprise, government] }
      responses:
        '201':
          description: Key created — plaintext shown once
          content:
            application/json:
              schema:
                type: object
                properties:
                  key:       { $ref: '#/components/schemas/ApiKey' }
                  plaintext: { type: string, example: "kv_live_ab12cd34..." }
...
  /webhooks:
    get:
      summary: List webhooks
      tags: [Webhooks]
      responses:
        '200':
          description: Webhook list
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhooks: { type: array, items: { type: object } }
    post:
      summary: Register a webhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:    { type: string, format: uri }
                events: { type: array, items: { type: string } }
      responses:
        '201':
          description: Webhook registered — secret shown once
...
  /tenants:
    post:
      summary: Create white-label tenant (super-admin)
      tags: [Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, name, owner_email, plan]
              properties:
                slug:          { type: string }
                name:          { type: string }
                owner_email:   { type: string, format: email }
                plan:          { type: string, enum: [starter, growth, scale, enterprise] }
                primary_color: { type: string, example: '#F97316' }
                app_name:      { type: string }
                default_language: { type: string, example: FR }
      responses:
        '201': { description: Tenant created }
  /tenants/{slug}/stats:
    get:
      summary: Get tenant usage stats
      tags: [Tenants]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User count, address count, API key count
...
  /packages:
    get:
      summary: List packages (admin/courier)
      tags: [Packages]
      responses:
        '200': { description: Package list }
    post:
      summary: Create a package
      tags: [Packages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sender_name, recipient_name, recipient_address, qr_code]
              properties:
                sender_name:       { type: string }
                recipient_name:    { type: string }
                recipient_address: { type: string }
                qr_code:           { type: string }
      responses:
        '201': { description: Package created }
...
  /inbox:
    get:
      summary: List inbox messages for authenticated user
      tags: [Inbox]
      parameters:
        - name: filter
          in: query
          schema: { type: string, enum: [all, unread, starred, archived] }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        '200': { description: Message list }
...